# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Vulnerability Assistance Platform API",
    "version": "1.0.0",
    "description": "API para plataforma de acompanhamento de pessoas em vulnerabilidade urbana",
    "contact": {
      "name": "API Support",
      "email": "support@vulnerability-platform.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3002",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT válido por 5 dias"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "required": [
          "nome",
          "email",
          "senha",
          "tipo_usuario"
        ],
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid",
            "description": "ID único do usuário (auto-gerado)",
            "readOnly": true
          },
          "nome": {
            "type": "string",
            "maxLength": 255,
            "description": "Nome completo do usuário"
          },
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 255,
            "description": "Email único do usuário"
          },
          "senha": {
            "type": "string",
            "description": "Senha do usuário (apenas para criação)"
          },
          "tipo_usuario": {
            "type": "string",
            "enum": [
              "missionario",
              "investidor"
            ],
            "description": "Tipo de usuário"
          },
          "status": {
            "type": "string",
            "enum": [
              "ativo",
              "inativo",
              "pendente"
            ],
            "description": "Status do usuário"
          },
          "data_cadastro": {
            "type": "string",
            "format": "date-time",
            "description": "Data de cadastro"
          },
          "ultimo_login": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Data do último login"
          }
        }
      },
      "Project": {
        "type": "object",
        "required": [
          "nome",
          "regiao",
          "missionary_owner_id"
        ],
        "properties": {
          "project_id": {
            "type": "string",
            "format": "uuid",
            "description": "ID único do projeto (auto-gerado)",
            "readOnly": true
          },
          "nome": {
            "type": "string",
            "maxLength": 255,
            "description": "Nome único do projeto"
          },
          "descricao": {
            "type": "string",
            "description": "Descrição do projeto"
          },
          "regiao": {
            "type": "string",
            "maxLength": 255,
            "description": "Região do projeto"
          },
          "data_criacao": {
            "type": "string",
            "format": "date-time",
            "description": "Data de criação"
          },
          "missionary_owner_id": {
            "type": "string",
            "format": "uuid",
            "description": "ID do missionário responsável (extraído do token JWT)"
          }
        }
      },
...
      "ApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Indica se a operação foi bem-sucedida"
          },
          "data": {
            "type": "object",
            "description": "Dados retornados"
          },
          "message": {
            "type": "string",
            "description": "Mensagem de resposta"
          },
          "count": {
            "type": "integer",
            "description": "Número de itens (para listas)"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "description": "Mensagem de erro"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/projects": {
      "post": {
        "summary": "Criar novo projeto",
        "tags": [
          "Projects"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Project"
              },
              "example": {
                "nome": "Projeto Vila Nova",
                "descricao": "Assistência a famílias em vulnerabilidade",
                "regiao": "Vila Nova"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Projeto criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Listar projetos",
        "tags": [
          "Projects"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "missionary_id",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Filtrar por ID do missionário (opcional - missionários veem apenas seus projetos)"
          },
          {
            "in": "query",
            "name": "regiao",
            "schema": {
              "type": "string"
            },
            "description": "Filtrar por região"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de projetos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{id}": {
      "get": {
        "summary": "Buscar projeto por ID",
        "tags": [
          "Projects"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID do projeto"
          }
        ],
        "responses": {
          "200": {
            "description": "Projeto encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "404": {
            "description": "Projeto não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Atualizar projeto",
        "tags": [
          "Projects"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID do projeto"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Project"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Projeto atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Deletar projeto",
        "tags": [
          "Projects"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID do projeto"
          }
        ],
        "responses": {
          "200": {
            "description": "Projeto deletado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "404": {
            "description": "Projeto não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/users": {
      "post": {
        "summary": "Criar novo usuário",
        "tags": [
          "Users"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              },
              "example": {
                "nome": "João Silva",
                "email": "joao@example.com",
                "senha": "123456",
                "tipo_usuario": "missionario"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuário criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Listar usuários",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "tipo_usuario",
            "schema": {
              "type": "string",
              "enum": [
                "missionario",
                "investidor"
              ]
            },
            "description": "Filtrar por tipo de usuário"
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "ativo",
                "inativo",
                "pendente"
              ]
            },
            "description": "Filtrar por status"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de usuários",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/{id}": {
      "get": {
        "summary": "Buscar usuário por ID",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID do usuário"
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "404": {
            "description": "Usuário não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Atualizar usuário",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID do usuário"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuário atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Usuário não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Deletar usuário",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID do usuário"
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário deletado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "404": {
            "description": "Usuário não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}